Rename SfCard to SfMobileCard
