--- packages/vue/src/components/molecules/SfCard/SfCard.vue
@@ -1,5 +1,0 @@
-<script src="./SfCard.js"></script>
-<template lang="html" src="./SfCard.html"></template>
-<style lang="scss">
-@import "~@storefront-ui/shared/styles/components/SfCard.scss";
-</style>
