+++ packages/vue/src/components/molecules/SfMobileCard/SfMobileCard.js
