+++ packages/vue/src/components/molecules/SfMobileCard/SfMobileCard.spec.ts
