+++ packages/vue/src/components/molecules/SfMobileCard/SfMobileCard.vue
@@ -1,0 +1,5 @@
+<script src="./SfMobileCard.js"></script>
+<template lang="html" src="./SfMobileCard.html"></template>
+<style lang="scss">
+@import "~@storefront-ui/shared/styles/components/SfMobileCard.scss";
+</style>
